% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{activate_backup}
\alias{activate_backup}
\title{Check IUCN backup}
\usage{
activate_backup(gbif.output, backup.iucn)
}
\arguments{
\item{gbif.output}{gbif.output object}

\item{backup.iucn}{backup.iucn object}
}
\value{
a \code{boolean}
}
\description{
Check whether IUCN Backup should be activated or not
}
