% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{backup_iucn}
\alias{backup_iucn}
\alias{backup_iucn-class}
\alias{set_backup_iucn}
\alias{show,backup_iucn-method}
\title{Parameters for IUCN backup}
\usage{
set_backup_iucn(
  do.backup,
  min.occurrence,
  min.threshold.effort,
  max.prop.own,
  max.prop.predator,
  max.prop.predator.outside,
  quantile.prop.predator
)

\S4method{show}{backup_iucn}(object)
}
\arguments{
\item{do.backup}{a \code{logical} whether backup was activated or not}

\item{min.occurrence}{a \code{integer}, the minimum number of gbif occurrences}

\item{min.threshold.effort}{a \code{integer}, the minimum sampling effort
threshold}

\item{max.prop.own}{a \code{numeric}, the maximum proportion of uncertain
cells within its own range}

\item{max.prop.predator}{a \code{numeric}, the maximum proportion of
uncertain cells within its predators range}

\item{max.prop.predator.outside}{a \code{numeric}, the maximum proportion of
uncertain cells outside its predators range}

\item{quantile.prop.predator}{a \code{numeric}, used to aggregate information
accross several predators.}

\item{object}{an object of class \code{backup_iucn}}
}
\description{
Class contained within a \code{\link{param_gbif}} object. A
  \code{iucn_backup} object contains the set of parameters used to 
  decide whether gbif data should be dropped to use iucn data instead:
  \itemize{
  \item whether backup was activated or not
  \item a minimum number of gbif occurrences 
  \item a minimum sampling effort threshold
  \item a maximum proportion of uncertain cells within its own range
  \item a maximum proportion of uncertain cells within the predator IUCN range 
  \item the quantile to aggregate proportion accross predator
  }
}
\section{Slots}{

\describe{
\item{\code{do.backup}}{a \code{logical} whether backup was activated or not}

\item{\code{min.occurrence}}{a \code{integer}, the minimum number of gbif occurrences}

\item{\code{min.threshold.effort}}{a \code{integer}, the minimum sampling effort
threshold}

\item{\code{max.prop.own}}{a \code{numeric}, the maximum proportion of uncertain
cells within its own range}

\item{\code{max.prop.predator}}{a \code{numeric}, the maximum proportion of
uncertain cells within its predators range}

\item{\code{max.prop.predator.outside}}{a \code{numeric}, the maximum proportion of
uncertain cells outside its predators range}

\item{\code{quantile.prop.predator}}{a \code{numeric}, used to aggregate information
accross several predators.}
}}

\author{
Remi Patin
}
