% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_tools.R
\name{buffer_iucn}
\alias{buffer_iucn}
\title{Buffer IUCN distribution}
\usage{
buffer_iucn(
  checklist,
  folder.iucn,
  folder.iucn.raster,
  data.mask,
  project.name,
  buffer.config,
  overwrite = TRUE,
  nb.cpu = 1
)
}
\arguments{
\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{folder.iucn}{a \code{character}: folders in which IUCN distributions
are stored}

\item{folder.iucn.raster}{a path to a folder with all IUCN range as raster.
used to determine which species have a IUCN range.}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{buffer.config}{a named \code{list} with taxonomic groups (Class, Order,
Family, Species) and associated distance that will be used to buffer IUCN
range. If several taxonomic groups are valid for a given species, the
lowest one will be prioritized. Buffer distances are given in km.}

\item{nb.cpu}{number of CPU to use}
}
\value{
a named list with the buffer associated to each species
}
\description{
\code{buffer_iucn} get the IUCN polygons for each class (old 
and new polygons), it then project to the \code{data.mask} projection,
buffer the polygons with \code{buffer.config} and rasterize the final projection.
}
\author{
Remi Patin
}
