% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_tools.R
\name{check_gbif_duplicata}
\alias{check_gbif_duplicata}
\title{Check gbif duplicated data}
\usage{
check_gbif_duplicata(checklist, folder.gbif, nb.cpu = 1)
}
\arguments{
\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}

\item{nb.cpu}{number of CPU to use}
}
\description{
\code{check_gbif_duplicata} retrieve the number of data in each
gbif file and return the file with the exact same number of data to 
identify gbif data that may be duplicated between species considered in gbif
as subspecies.
}
\author{
Remi Patin
}
