% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{check_name_config}
\alias{check_name_config}
\title{Check if a name is present in a sampling effort configuration}
\usage{
check_name_config(sampling.effort.config, thisname)
}
\arguments{
\item{sampling.effort.config}{a named \code{list} with one element per effort
layer to be calculated. Contains layer name as well as their associated taxa.}

\item{thisname}{name to be checked}
}
\value{
vector
}
\description{
Check if a name is present in a sampling effort configuration
}
\keyword{internal}
