% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{check_taxa}
\alias{check_taxa}
\title{Check if a taxonomic group is present in the checklist}
\usage{
check_taxa(this.taxa, checklist)
}
\arguments{
\item{this.taxa}{\code{character} taxonomic group to be checked}

\item{checklist}{a \code{data.frame} with information on all species of interest}
}
\value{
a boolean
}
\description{
Check if a taxonomic group is present in the checklist
}
\keyword{internal}
