% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.check_checklist}
\alias{.check_checklist}
\title{Check checklist data.frame}
\usage{
.check_checklist(x)
}
\arguments{
\item{x}{data.frame to be checked}
}
\value{
a boolean
}
\description{
Check whether a data.frame is a valid 
checklist
}
\keyword{internal}
