% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.check_metaweb}
\alias{.check_metaweb}
\title{Check metaweb data.frame}
\usage{
.check_metaweb(x)
}
\arguments{
\item{x}{data.frame to be checked}
}
\value{
a boolean
}
\description{
Check whether a data.frame is a valid metaweb
}
\keyword{internal}
