% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.check_metaweb_checklist}
\alias{.check_metaweb_checklist}
\title{Check metaweb and checklist compatibility}
\usage{
.check_metaweb_checklist(metaweb, checklist)
}
\arguments{
\item{metaweb}{a \code{data.frame} with all species interactions listed}

\item{checklist}{a \code{data.frame} with information on all species of interest}
}
\value{
a boolean
}
\description{
Check that checklist and metaweb have the same code and 
species name
}
