% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.check_trophic.groups}
\alias{.check_trophic.groups}
\title{Check trophic.groups data.frame}
\usage{
.check_trophic.groups(x)
}
\arguments{
\item{x}{data.frame to be checked}
}
\value{
a boolean
}
\description{
Check whether a data.frame is a valid 
trophic.groups
}
\keyword{internal}
