% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.filter_df_by_checklist}
\alias{.filter_df_by_checklist}
\title{filter metaweb to keep only valid species code}
\usage{
.filter_df_by_checklist(df, checklist, col.name, project.name)
}
\arguments{
\item{df}{a data.frame to be checked}

\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{col.name}{a character determining the column name to check}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}
}
\value{
a data.frame
}
\description{
Filter metaweb to keep only valid species code
}
\keyword{internal}
