% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testdfcolnames}
\alias{.fun_testdfcolnames}
\title{Check data.frame column names}
\usage{
.fun_testdfcolnames(x, thiscolnames)
}
\arguments{
\item{x}{object to be checked}

\item{thiscolnames}{a set of column nmaes}
}
\value{
a boolean
}
\description{
Check whether a data.frame have valid column names
}
\keyword{internal}
