% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIf01}
\alias{.fun_testIf01}
\title{Check if object is included in 0-1}
\usage{
.fun_testIf01(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
a boolean
}
\description{
Check if object is included in 0-1
}
\keyword{internal}
