% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfDirExists}
\alias{.fun_testIfDirExists}
\title{Check folder existence}
\usage{
.fun_testIfDirExists(x)
}
\arguments{
\item{x}{folder to check}
}
\value{
a boolean
}
\description{
Check whether a folder exists
}
\keyword{internal}
