% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfIn}
\alias{.fun_testIfIn}
\title{Check object values}
\usage{
.fun_testIfIn(x, values)
}
\arguments{
\item{x}{object to be checked}

\item{values}{a set of values}
}
\value{
a boolean
}
\description{
Check whether an object only contains a set of values
}
\keyword{internal}
