% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfInherits}
\alias{.fun_testIfInherits}
\title{Check class of an object}
\usage{
.fun_testIfInherits(x, thisclass)
}
\arguments{
\item{x}{object to be checked}

\item{thisclass}{a class to check}
}
\value{
a boolean
}
\description{
Check whether an object inherits a given class
}
\keyword{internal}
