% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfNum}
\alias{.fun_testIfNum}
\title{Check if object is a numeric}
\usage{
.fun_testIfNum(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
a boolean
}
\description{
Check whether an object is a numeric
}
\keyword{internal}
