% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfPosInt}
\alias{.fun_testIfPosInt}
\title{Check if object is a positive integer}
\usage{
.fun_testIfPosInt(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
a boolean
}
\description{
Check if object is a positive integer
}
\keyword{internal}
