% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.fun_testIfPosNum}
\alias{.fun_testIfPosNum}
\title{Check if object is a positive numeric}
\usage{
.fun_testIfPosNum(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
a boolean
}
\description{
Check if object is a positive numeric
}
\keyword{internal}
