% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.register_cluster}
\alias{.register_cluster}
\title{Register a cluster if need be}
\usage{
.register_cluster(nb.cpu)
}
\arguments{
\item{nb.cpu}{number of CPU to use}
}
\description{
Register a cluster to be used with \code{foreach}
}
\keyword{internal}
