% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_assemble_functions.R
\name{extract_gbif}
\alias{extract_gbif}
\title{Extract gbif occurrences for a single species}
\usage{
extract_gbif(
  this.species,
  this.code,
  param.gbif,
  occurrence.dir,
  project.name,
  logfile
)
}
\arguments{
\item{this.species}{a \code{character}: species name}

\item{this.code}{a \code{character}: species code}

\item{param.gbif}{a \code{param_gbif} gbif configuration}

\item{occurrence.dir}{a \code{character}, folder to store occurrence data}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{logfile}{a \code{character}, file to store logs}
}
\description{
\code{extract_gbif} retrieve gbif data from files with a given 
configuration
}
\author{
Remi Patin
}
