% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{find_iucn_index}
\alias{find_iucn_index}
\title{Find a species in IUCN range data.frame}
\usage{
find_iucn_index(df, this.species, species_col, capitalized)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{this.species}{a \code{character}, species to be found}

\item{species_col}{a \code{character}, column name for the species name}

\item{capitalized}{\code{boolean}, TRUE if column names are capitalized}
}
\value{
a boolean
}
\description{
Find a species in IUCN range data.frame
}
\keyword{internal}
