% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{get_occurrence_summary}
\alias{get_occurrence_summary}
\title{Get occurrence summary}
\usage{
get_occurrence_summary(this.df)
}
\arguments{
\item{this.df}{a data.frame}
}
\value{
a named list
}
\description{
Get occurrence summary
}
