% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{get_predator_summary}
\alias{get_predator_summary}
\title{Get summary of predator occurrences}
\usage{
get_predator_summary(this.trophic)
}
\arguments{
\item{this.trophic}{a \code{data.frame} with occurrence data: columns cell,
x, y, presence, inside_iucn, SpeciesName, Code and all prey occurrences as 
columns}
}
\value{
a data.frame
}
\description{
Get prevalence and summary of occurrences inside and outside
IUCN range for a given predator.
}
