% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_tools.R
\name{get_species_buffer}
\alias{get_species_buffer}
\title{get species-buffer association}
\usage{
get_species_buffer(buffer.config, checklist)
}
\arguments{
\item{buffer.config}{a named \code{list} with taxonomic groups (Class, Order,
Family, Species) and associated distance that will be used to buffer IUCN
range. If several taxonomic groups are valid for a given species, the
lowest one will be prioritized. Buffer distances are given in km.}

\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{folder.iucn.raster}{a path to a folder with all IUCN range as raster.
used to determine which species have a IUCN range.}
}
\value{
a named list with the buffer associated to each species
}
\description{
\code{get_species_buffer} get the buffer corresponding to each
  species, based on a list with buffer distance based on taxonomy
}
\author{
Remi Patin
}
