% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate ggplot2 color scale}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
character
}
\description{
Emulate ggplot2 color scale
}
\keyword{internal}
