% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\title{Load trophic dataset}
\usage{
load_dataset(project.name)
}
\arguments{
\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}
}
\value{
a \code{\link{trophic_dataset}}
}
\description{
\code{load_dataset} loads a trophic dataset
}
