% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_gbif_data}
\alias{load_gbif_data}
\title{Load gbif data}
\usage{
load_gbif_data(species.name, folder.gbif, filter.atlas, all.columns = FALSE)
}
\arguments{
\item{species.name}{a \code{character} corresponding to the species name}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}
}
\value{
a data.frame or a character if the loading failed
}
\description{
\code{load_gbif_data} find data for a given species 
within \code{folder.gbif}. Data is then read as a data.frame
}
