% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{locate_iucn_distribution}
\alias{locate_iucn_distribution}
\title{Locate IUCN distribution}
\usage{
locate_iucn_distribution(species.code, folder.iucn)
}
\arguments{
\item{species.code}{a \code{character} corresponding to the species code}

\item{folder.iucn}{a \code{character}: folders in which IUCN distributions
are stored}
}
\value{
a data.frame or a character if the loading failed
}
\description{
\code{locate_iucn_distribution} find distribution for a given
 species within \code{folder.iucn}. Path to file is then returned
}
