% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_border.R
\name{map_filtered_prey}
\alias{map_filtered_prey}
\title{Information on cells filtered due to prey}
\usage{
map_filtered_prey(project.name, vector_data, data.mask, suffix)
}
\arguments{
\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{vector_data}{a \code{SpatVector} with administrative boundary}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{suffix}{a \code{character} with the name of the administrative boundary}
}
\description{
\code{map_filtered_prey} represent the mean probability of
filtering a cell due to uncertain prey with different aggregation levels
}
\author{
Remi Patin
}
