% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_border.R
\name{map_uncertain}
\alias{map_uncertain}
\title{Plot absence uncertainty}
\usage{
map_uncertain(project.name, vector_data, data.mask, suffix)
}
\arguments{
\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{vector_data}{a \code{SpatVector} with administrative boundary}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{suffix}{a \code{character} with the name of the administrative boundary}
}
\description{
\code{map_uncertain} represent the mean probability for an absence
to be classified as uncertain
}
\author{
Remi Patin
}
