% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{occurrences_as_raster}
\alias{occurrences_as_raster}
\title{Write occurrences and data status as raster}
\usage{
occurrences_as_raster(
  occurrence.file,
  data.mask,
  occurrence.rast.dir,
  status.rast.dir,
  species,
  full.run
)
}
\arguments{
\item{occurrence.file}{file with species}

\item{data.mask}{a \code{SpatRaster}, mask used for rasterization}

\item{occurrence.rast.dir}{a \code{character}, folder to store occurrence 
raster files.}

\item{status.rast.dir}{a \code{character}, folder to store data status 
raster files.}

\item{species}{a \code{character}, code of current species.}

\item{full.run}{a \code{boolean}, whether raster should be saved or only 
filenames should be returned}
}
\value{
a \code{list} with links to status and occurrence raster file
}
\description{
Write occurrences and data status as raster
}
