% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{param_gbif}
\alias{param_gbif}
\alias{param_gbif-class}
\alias{set_param_gbif}
\alias{show,param_gbif-method}
\title{Parameters for gbif data extraction}
\usage{
set_param_gbif(
  use.gbif,
  folder.gbif,
  min.gbif,
  filter.atlas,
  buffer.config,
  folder.iucn.buffer,
  sampling.effort,
  quantile.absence.certain,
  prop.prey.certain,
  uncertain.value,
  backup.iucn,
  ...
)

\S4method{show}{param_gbif}(object)
}
\arguments{
\item{use.gbif}{a \code{logical}, whether gbif was used or not}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}

\item{min.gbif}{a \code{integer}, the minimum number of gbif occurrences}

\item{filter.atlas}{a \code{booelan}, whether atlas data should be filtered
out (through column \code{coordinateUncertaintyInMeters})}

\item{buffer.config}{a \code{list}, the configuration for iucn buffer}

\item{folder.iucn.buffer}{a \code{character}, folder in which IUCN range
are stored as raster files.}

\item{sampling.effort}{a \code{sampling_effort} object describing the sampling
effort associated to all taxa.}

\item{quantile.absence.certain}{a \code{numeric} (\emph{default}
\code{0.1}), between 0 and 1. The quantile of sampling effort below which
a species absence is considered as uncertain}

\item{prop.prey.certain}{a \code{numeric} (\emph{default} \code{0.8}),
between 0 and 1. The minimum proportion of prey occurrence considered as
certain for a raster cell to be included in the species dataset.}

\item{uncertain.value}{a \code{integer} (\code{0} or \code{1} ;
\emph{default} \code{0}). The value given to uncertain absences in cells
where more than \code{prop.prey.certain} prey are considered as certain.}

\item{backup.iucn}{a \code{\link{backup_iucn}} object}

\item{...}{additionnal parameters given to \code{\link{set_backup_iucn}}}

\item{object}{an object of class \code{param_gbif}}
}
\description{
Class contained within a \code{\link{param_trophic}} object. A
  \code{param_gbif} object contains the set of parameters used to 
  extract gbif occurrences. It contains:
 \itemize{
  \item whether gbif data were ever used
  \item the folder with gbif data
  \item the minimum number of gbif occurrences 
  \item whether atlas data should be filtered out (through column 
  \code{coordinateUncertaintyInMeters})
  \item The configuration for iucn buffer
  \item The folder with buffered IUCN ranges
  \item The quantile of sampling effort below which a species absence is
  considered as uncertain
  \item The minimum proportion of prey occurrence considered as certain for
  a raster cell to be included in the species dataset.
  \item The value given to uncertain absences in cells
  where more than \code{prop.prey.certain} prey are considered as certain.
  \item The parameter for IUCN backup (whether IUCN data should be used when
  gbif data are not sufficient)
  }
}
\section{Slots}{

\describe{
\item{\code{use.gbif}}{a \code{logical}, whether gbif was used or not}

\item{\code{folder.gbif}}{a \code{character}: folders in which gbif data are stored}

\item{\code{min.gbif}}{a \code{integer}, the minimum number of gbif occurrences}

\item{\code{filter.atlas}}{a \code{booelan}, whether atlas data should be filtered
out (through column \code{coordinateUncertaintyInMeters})}

\item{\code{buffer.config}}{a \code{list}, the configuration for iucn buffer}

\item{\code{species.buffer}}{a named \code{list} with the buffer associated to each species}

\item{\code{folder.iucn.buffer}}{a \code{character}, folder in which IUCN range
are stored as raster files.}

\item{\code{quantile.absence.certain}}{a \code{numeric} (\emph{default}
\code{0.1}), between 0 and 1. The quantile of sampling effort below which
a species absence is considered as uncertain}

\item{\code{prop.prey.certain}}{a \code{numeric} (\emph{default} \code{0.8}),
between 0 and 1. The minimum proportion of prey occurrence considered as
certain for a raster cell to be included in the species dataset.}

\item{\code{uncertain.value}}{a \code{integer} (\code{0} or \code{1} ;
\emph{default} \code{0}). The value given to uncertain absences in cells
where more than \code{prop.prey.certain} prey are considered as certain.}

\item{\code{backup.iucn}}{a \code{\link{backup_iucn}} object}

\item{\code{sampling.effort}}{a \code{sampling_effort} object describing the sampling
effort associated to all taxa.}
}}

\author{
Remi Patin
}
