% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{param_subsampling}
\alias{param_subsampling}
\alias{param_subsampling-class}
\alias{set_param_subsampling}
\alias{show,param_subsampling-method}
\title{Parameters for subsampling dataset}
\usage{
set_param_subsampling(
  method,
  max.presences,
  max.absences.inside,
  min.absences.outside,
  max.absences.outside,
  prop.outside
)

\S4method{show}{param_subsampling}(object)
}
\arguments{
\item{method}{a \code{logical}, whether gbif was used or not}

\item{max.presences}{a \code{integer} maximum number of presences}

\item{max.absences.inside}{a \code{integer} maximum number of absences inside
IUCN range}

\item{min.absences.outside}{a \code{integer} minimum number of absences
outside IUCN range}

\item{max.absences.outside}{a \code{integer} maximum number of absences
outside IUCN range}

\item{prop.outside}{a \code{numeric}, target proportion of
absences outside IUCN range (relative to total data within IUCN range)}

\item{object}{an object of class \code{param_subsampling}}
}
\description{
Class contained within a \code{\link{param_trophic}} object. A
  \code{param_subsampling} object contains the set of parameters used to 
  extract subsample occurrences in the final dataset. It contains:
 \itemize{
  \item the method used (only random for now)
  \item the maximum number of presences
  \item the maximum number of absences inside IUCN range
  \item the minimum number of absences outside IUCN range (if applicable)
  \item the maximum number of absences outside IUCN range (if applicable)
  \item the target proportion of absences outside IUCN range (relative to 
  total data within IUCN range)
  }
}
\section{Slots}{

\describe{
\item{\code{method}}{a \code{logical}, whether gbif was used or not}

\item{\code{max.presences}}{a \code{integer} maximum number of presences}

\item{\code{max.absences.inside}}{a \code{integer} maximum number of absences inside
IUCN range}

\item{\code{min.absences.outside}}{a \code{integer} minimum number of absences
outside IUCN range}

\item{\code{max.absences.outside}}{a \code{integer} maximum number of absences
outside IUCN range}

\item{\code{prop.outside}}{a \code{numeric}, target proportion of
absences outside IUCN range (relative to total data within IUCN range)}
}}

\author{
Remi Patin
}
