% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_tools.R
\name{rasterize_iucn}
\alias{rasterize_iucn}
\title{Buffer IUCN distribution}
\usage{
rasterize_iucn(checklist, folder.iucn, data.mask, project.name, nb.cpu = 1)
}
\arguments{
\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{folder.iucn}{a \code{character}: folders in which IUCN distributions
are stored}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{nb.cpu}{number of CPU to use}
}
\description{
\code{buffer_iucn} get the IUCN polygons for each class (old 
and new polygons), it then project to the \code{data.mask} projection,
buffer the polygons with \code{buffer.config} and rasterize the final projection.
}
\author{
Remi Patin
}
