% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{set_folder}
\alias{set_folder}
\title{Set folder for prepare_dataset}
\usage{
set_folder(project.name)
}
\arguments{
\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}
}
\value{
a named list
}
\description{
Initialize and set folder name for \code{\link{prepare_dataset}}
}
