% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{subsample_dataset}
\alias{subsample_dataset}
\title{Subsample dataset}
\usage{
subsample_dataset(df, param.subsampling)
}
\arguments{
\item{df}{a data.frame}

\item{param.subsampling}{a \code{param_subsampling} object}
}
\value{
a named list
}
\description{
subsample a dataset
}
