% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_taxonomic_conflict.R,
%   R/class_trophic_conflict.R
\docType{class}
\name{taxonomic_conflict}
\alias{taxonomic_conflict}
\alias{taxonomic_conflict-class}
\alias{detect_taxonomic_conflict}
\alias{show,taxonomic_conflict-method}
\alias{summary_conflict}
\alias{plot,taxonomic_conflict,missing-method}
\title{Object class describing taxonomic conflicts within trophic groups}
\usage{
detect_taxonomic_conflict(metaweb, trophic.groups, checklist, project.name)

\S4method{show}{taxonomic_conflict}(object)

summary_conflict(object, type = "Class")

\S4method{plot}{taxonomic_conflict,missing}(x, type = "Class")

detect_taxonomic_conflict(metaweb, trophic.groups, checklist, project.name)

\S4method{show}{taxonomic_conflict}(object)

summary_conflict(object, type = "Class")

\S4method{plot}{taxonomic_conflict,missing}(x, type = "Class")
}
\arguments{
\item{metaweb}{a \code{data.frame} with all species interactions listed}

\item{trophic.groups}{a \code{data.frame} assigning trophic groups to all 
species of interest}

\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{type}{a character, either 'Class', 'Order' or 'Family' determining
the level of conflict that needs to be summarised}
}
\description{
Class returned by \code{\link{detect_taxonomic_conflict}}, 
that list species within the same trophic group with different taxonomy but 
the same predator

Class returned by \code{\link{detect_taxonomic_conflict}}, 
that list species within the same trophic group with different taxonomy but 
the same predator
}
\section{Slots}{

\describe{
\item{\code{class.summary}}{a \code{data.frame} summarising the number of class-levels conflicts}

\item{\code{class.conflict}}{a \code{data.frame} with all row from \code{metaweb} with 
class-level conflicts}

\item{\code{order.summary}}{a \code{data.frame} summarising the number of order-levels conflicts}

\item{\code{order.conflict}}{a \code{data.frame} with all row from \code{metaweb} with 
order-level conflicts}

\item{\code{metaweb}}{a \code{data.frame} with all species interactions listed}

\item{\code{trophic.groups}}{a \code{data.frame} with information on trophic groups}

\item{\code{checklist}}{a \code{data.frame} with information on all species of interest}

\item{\code{project.name}}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{\code{class.summary}}{a \code{data.frame} summarising the number of class-levels conflicts}

\item{\code{class.conflict}}{a \code{data.frame} with all row from \code{metaweb} with 
class-level conflicts}

\item{\code{order.summary}}{a \code{data.frame} summarising the number of order-levels conflicts}

\item{\code{order.conflict}}{a \code{data.frame} with all row from \code{metaweb} with 
order-level conflicts}

\item{\code{metaweb}}{a \code{data.frame} with all species interactions listed}

\item{\code{trophic.groups}}{a \code{data.frame} with information on trophic groups}

\item{\code{checklist}}{a \code{data.frame} with information on all species of interest}

\item{\code{project.name}}{a \code{character} indicating the folder in which logfiles 
and data may be written}
}}

\examples{

showClass("taxonomic_conflict")

showClass("taxonomic_conflict")
}
\author{
Remi Patin
}
