% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{trophic_dataset}
\alias{trophic_dataset}
\alias{trophic_dataset-class}
\alias{prepare_dataset_gbif}
\alias{show,trophic_dataset-method}
\alias{.filter_dataset.check.args}
\title{Dataset for running trophic SDM}
\usage{
prepare_dataset_gbif(
  checklist,
  metaweb,
  folder.gbif,
  folder.iucn,
  sampling.effort,
  data.mask,
  project.name,
  min.gbif,
  buffer.iucn,
  quantile.absence.certain,
  prop.prey.certain,
  uncertain.value,
  nb.cpu,
  ...
)

\S4method{show}{trophic_dataset}(object)

.filter_dataset.check.args(
  min.presence,
  min.absence,
  subsample.min.prevalence,
  subsample.max.absence,
  min.prevalence.prey,
  min.absence.prey
)
}
\arguments{
\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{metaweb}{a \code{data.frame} with all species interactions listed}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}

\item{folder.iucn}{a \code{character}: folders in which IUCN distributions
are stored}

\item{sampling.effort}{a \code{\link{sampling_effort}} object, generated by
\code{\link{calc_sampling_effort}} containing several layers of sampling
effort based on taxonomic group and the association between species and
the sampling effort layer}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{min.gbif}{a \code{numeric} (\emph{default} \code{25}), the minimum
number of gbif occurrences below which a species is filtered.}

\item{buffer.iucn}{a named \code{list} with taxonomic groups (Class, Order,
Family, Species) and associated distance that will be used to buffer IUCN
range. If several taxonomic groups are valid for a given species, the
lowest one will be prioritized.}

\item{quantile.absence.certain}{a \code{numeric} (\emph{default}
\code{0.1}), between 0 and 1. The quantile of sampling effort below which
a species absence is considered as uncertain}

\item{prop.prey.certain}{a \code{numeric} (\emph{default} \code{0.8}),
between 0 and 1. The minimum proportion of prey occurrence considered as
certain for a raster cell to be included in the species dataset.}

\item{uncertain.value}{a \code{integer} (\code{0} or \code{1} ;
\emph{default} \code{0}). The value given to uncertain absences in cells
where more than \code{prop.prey.certain} prey are considered as certain.}

\item{nb.cpu}{number of CPU to use}

\item{...}{additional parameter given to \code{\link{filter_dataset}} to
filter the dataset at the end of the workflow.}

\item{object}{an object of class \code{trophic_dataset}}
}
\description{
Class returned by \code{\link{prepare_dataset_gbif}} and
  \code{\link{prepare_dataset_iucn}}, that prepare a dataset for a trophic
  SDM, based on a gbif or IUCN workflow.
}
\details{
\describe{
  \item{gbif dataset (\code{prepare_gbif_dataset})}{ Workflow to prepare gbif
  dataset}
  \item{IUCN dataset (\code{prepare_iucn_dataset})}{ Workflow to prepare iucn
  dataset}
  }
}
\section{Slots}{

\describe{
\item{\code{summary.species}}{a \code{data.frame} with the list of species and 
the total number of presence and absence as well as additional information}

\item{\code{file.link}}{a named \code{list} with a path to data for each species}

\item{\code{metaweb}}{a \code{data.frame} with all species interactions listed}

\item{\code{checklist}}{a \code{data.frame} with information on all species of
interest}

\item{\code{kept.species}}{a \code{character} vector with all species kept}

\item{\code{filtered.species}}{a named \code{list} vector with all species removed 
as well as the motivation of removal}

\item{\code{param}}{a \code{list} with all parameters used to generate the dataset}

\item{\code{datatype}}{a \code{character} determining the origin of the dataset:
either 'gbif' or 'iucn'.}

\item{\code{project.name}}{a \code{character} indicating the folder in which
logfiles and data may be written}

\item{\code{data.mask}}{a \code{SpatRaster} with the extent and the grid used 
in the project}
}}

\examples{

showClass("trophic_dataset")
}
\author{
Remi Patin
}
