% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{trophic_files}
\alias{trophic_files}
\alias{trophic_files-class}
\alias{show,trophic_files-method}
\title{Set of links to data files}
\usage{
\S4method{show}{trophic_files}(object)
}
\arguments{
\item{object}{an object of class \code{trophic_files}}
}
\description{
Class contained within \code{\link{trophic_dataset}}. A
  \code{trophic_files} object contains a set of file links to the different
  dataset generated throughout the function \code{\link{prepare_dataset}}:
  \itemize{
  \item Occurrence files (focal species only)
  \item Occurrence files (as \code{\link{SpatRaster}})
  \item Status files (as \code{\link{SpatRaster}})
  \item Raw trophic files (focal species and its prey)
  \item Final trophic files (final subsampled dataset)
  }
}
\section{Slots}{

\describe{
\item{\code{occurrence}}{a named \code{vector} with a path to raw 
occurrence data for each species.}

\item{\code{occurrence.rast}}{a named \code{vector} with a path to raw raster 
occurrence data for each species.}

\item{\code{status.rast}}{a named \code{vector} with a path to raster data 
containing status of data (certain vs uncertain) for each species.}

\item{\code{trophic}}{a named \code{vector} with a path to filtered
trophic data for each species.}

\item{\code{trophic.raw}}{a named \code{vector} with a path to raw
trophic data for each species.}
}}

\author{
Remi Patin
}
