% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R
\docType{class}
\name{trophic_summary}
\alias{trophic_summary}
\alias{trophic_summary-class}
\alias{show,trophic_summary-method}
\title{Summary of trophic dataset}
\usage{
\S4method{show}{trophic_summary}(object)
}
\arguments{
\item{object}{an object of class \code{trophic_summary}}
}
\description{
Class contained within \code{\link{trophic_dataset}}. A
  \code{trophic_summary} object contains a set of summary \code{data.frame} 
  for the \code{\link{trophic_dataset}} object:
  \itemize{
  \item occurrence summary
  \item final dataset summary
  \item prey summary
  \item filtered species summary
  \item protected species summary
  \item patrimonial species summary
  }
}
\section{Slots}{

\describe{
\item{\code{occurrence}}{a \code{data.frame} with the list of species and 
the total number of presence and absence in the occurrence dataset extracted
from gbif and independently of prey cell selection.}

\item{\code{trophic}}{a \code{data.frame} with a summary of the actual 
trophic dataset, i.e. list of species and the total number of presence 
and absence inside/outside IUCN range as well as additional information 
such as the number of prey.}

\item{\code{prey}}{a \code{data.frame} with a summary of occurrence and 
prevalence for all prey associated to a predator dataset.}

\item{\code{filtered}}{a \code{data.frame} with a summary of species that were 
filtered throughout the data generation}

\item{\code{protected}}{a \code{data.frame} with a summary of species protected by 
the EU habitat directive}

\item{\code{patrimonial}}{a \code{data.frame} with a summary of species listed as 
patrimonial species by the french INPN}
}}

\author{
Remi Patin
}
