% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tools.R
\name{.write_logfile}
\alias{.write_logfile}
\title{Write a logfile}
\usage{
.write_logfile(out.log, logfile, project.name, open = "w", silent = FALSE)
}
\arguments{
\item{out.log}{character to be written to the logfile}

\item{logfile}{file name}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{open}{type of opening ('w' for standard, 'a' for appending)}
}
\description{
Write a logfile and create appropriate folders if needbe
}
\keyword{internal}
