% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_assemble_functions.R
\name{extract_iucn}
\alias{extract_iucn}
\alias{assemble_trophic}
\title{Extract IUCN occurrences for a single species}
\usage{
extract_iucn(
  this.species,
  this.code,
  folder.iucn,
  occurrence.dir,
  project.name,
  logfile
)

assemble_trophic(this.species, listcode, file.occurrence.link, raw.dir, param)
}
\arguments{
\item{this.species}{a \code{character}: species name}

\item{this.code}{a \code{character}: species code}

\item{folder.iucn}{a \code{character}, folder with IUCN distribution as raster}

\item{occurrence.dir}{a \code{character}, folder to store occurrence data}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{logfile}{a \code{character}, file to store logs}
}
\description{
\code{extract_iucn} retrieve IUCN data from files with a given 
configuration

\code{extract_iucn} retrieve IUCN data from files with a given 
configuration
}
\author{
Remi Patin
}
