% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_assemble_functions.R
\name{extract_iucn}
\alias{extract_iucn}
\title{Extract IUCN occurrences for a single species}
\usage{
extract_iucn(
  this.species,
  this.code,
  data.mask = data.mask,
  folder.iucn,
  occurrence.dir,
  occurrence.rast.dir,
  status.rast.dir,
  project.name,
  logfile,
  overwrite
)
}
\arguments{
\item{this.species}{a \code{character}: species name}

\item{this.code}{a \code{character}: species code}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{folder.iucn}{a \code{character}, folder with IUCN distribution as raster}

\item{occurrence.dir}{a \code{character}, folder to store occurrence data}

\item{occurrence.rast.dir}{a \code{character}, folder to store occurrence 
raster files.}

\item{status.rast.dir}{a \code{character}, folder to store data status 
raster files.}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{logfile}{a \code{character}, file to store logs}

\item{overwrite}{a \code{boolean}, whether occurrence dataset should be 
overwritten.}
}
\description{
\code{extract_iucn} retrieve IUCN data from files with a given 
configuration
}
\author{
Remi Patin
}
