% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gbif_outsider.R
\docType{class}
\name{gbif_outsider}
\alias{gbif_outsider}
\alias{gbif_outsider-class}
\alias{detect_gbif_outsider}
\alias{show,gbif_outsider-method}
\alias{summary_outsider}
\alias{plot,gbif_outsider,missing-method}
\title{Object class containing gbif observation outside of IUCN range}
\usage{
detect_gbif_outsider(
  checklist,
  folder.gbif,
  folder.iucn,
  project.name,
  nb.cpu = 1
)

\S4method{show}{gbif_outsider}(object)

summary_outsider(object, type = "Class")

\S4method{plot}{gbif_outsider,missing}(x, type = "Class", nb.cpu = 1, plot.folder = "gbif_outsider")
}
\arguments{
\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}

\item{folder.iucn}{a \code{character}: folders in which IUCN distributions
are stored}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{object}{an object of class \code{gbif_outsider}}

\item{type}{a character, either 'Class', 'Order' or 'Family' determining
the level of conflict that needs to be summarised}

\item{plot.folder}{a path to store plot results when \code{type = "Species"}}
}
\description{
Class returned by \code{\link{detect_gbif_outsider}}, 
that list gbif observations outside of IUCN range. Also provides tools to 
explore the distribution of outsiders.
}
\section{Slots}{

\describe{
\item{\code{outsider}}{a \code{data.frame} with all gbif outsider}

\item{\code{data.summary}}{a \code{data.frame} summarising the total number of gbif 
data for each species}

\item{\code{iucn.distrib}}{a \code{list} with path for iucn distribution for each species}

\item{\code{checklist}}{a \code{data.frame} with information on all species of interest}

\item{\code{project.name}}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{\code{species.failed}}{a \code{character} listing all species for which no 
data were found}
}}

\examples{

showClass("gbif_outsider")
}
\author{
Remi Patin
}
