% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R,
%   R/prepare_dataset_gbif.R
\docType{class}
\name{trophic_dataset}
\alias{trophic_dataset}
\alias{trophic_dataset-class}
\alias{show,trophic_dataset-method}
\alias{filter_species}
\alias{filter_species,trophic_dataset-method}
\alias{subsample}
\alias{subsample,trophic_dataset-method}
\alias{filter_prey}
\alias{filter_prey,trophic_dataset-method}
\alias{load_data}
\alias{load_data,trophic_dataset-method}
\alias{reset}
\alias{reset,trophic_dataset-method}
\alias{summary_trophic}
\alias{summary_trophic,trophic_dataset-method}
\alias{plot_uncertain}
\alias{plot_uncertain,trophic_dataset-method}
\alias{plot_trophic}
\alias{plot_trophic,trophic_dataset-method}
\alias{rasterize_uncertain}
\alias{rasterize_uncertain,trophic_dataset-method}
\alias{rasterize_prey_filtering}
\alias{rasterize_prey_filtering,trophic_dataset-method}
\alias{prepare_dataset_gbif}
\title{Dataset for running trophic SDM}
\usage{
\S4method{show}{trophic_dataset}(object)

filter_species(x, ...)

\S4method{filter_species}{trophic_dataset}(x, min.presence, min.absence)

subsample(x, ...)

\S4method{subsample}{trophic_dataset}(x, subsample.min.prevalence, subsample.max.absence)

filter_prey(x, ...)

\S4method{filter_prey}{trophic_dataset}(x, min.prevalence.prey, min.absence.prey)

load_data(x, ...)

\S4method{load_data}{trophic_dataset}(x, SpeciesName, Code, type = "trophic")

reset(x, ...)

\S4method{reset}{trophic_dataset}(x)

summary_trophic(x, ...)

\S4method{summary_trophic}{trophic_dataset}(x, info = "trophic_dataset")

plot_uncertain(x, ...)

\S4method{plot_uncertain}{trophic_dataset}(x, nb.cpu = 1)

plot_trophic(x, ...)

\S4method{plot_trophic}{trophic_dataset}(x, nb.cpu = 1)

rasterize_uncertain(x, ...)

\S4method{rasterize_uncertain}{trophic_dataset}(x, raster.by = "Class")

rasterize_prey_filtering(x, ...)

\S4method{rasterize_prey_filtering}{trophic_dataset}(x, raster.by = "Class")

prepare_dataset_gbif(
  checklist,
  metaweb,
  folder.gbif,
  folder.iucn.buffer,
  sampling.effort,
  data.mask,
  project.name,
  min.gbif,
  buffer.config,
  quantile.absence.certain,
  prop.prey.certain,
  uncertain.value,
  nb.cpu,
  ...
)
}
\arguments{
\item{object}{an object of class \code{trophic_dataset}}

\item{x}{an object of class \code{trophic_dataset}}

\item{...}{additional parameter given to \code{\link{filter_dataset}} to
filter the dataset at the end of the workflow.}

\item{min.presence}{a \code{integer} used as a filtering parameter. Any 
species with less presence than \code{min.presence} in its trophic dataset
is removed.}

\item{min.absence}{a \code{integer} used as a filtering parameter. Any 
species with less absence than \code{min.absence} in its trophic dataset
is removed.}

\item{subsample.min.prevalence}{a \code{numeric} used as a subsampling
parameter. Used to subsample absences in trophic dataset with prevalence below 
\code{subsample.min.prevalence}.}

\item{subsample.max.absence}{a \code{numeric} used as a subsampling
parameter. Used to subsample absences in trophic dataset with more absences 
than \code{subsample.max.absence}.}

\item{min.prevalence.prey}{a \code{numeric} used as a filtering parameter. Any 
prey  with a prevalence in predator presence cell below 
\code{min.prevalence.prey}  is removed from the predator dataset.}

\item{min.absence.prey}{a \code{integer} used as a filtering parameter. Any 
prey less absence than \code{min.absence.prey} is removed from the predator
 dataset.}

\item{SpeciesName}{a \code{character}, a species name to retrieve data or 
information}

\item{Code}{a \code{character}, a species code to retrieve data or information}

\item{type}{a \code{character}, the type of data to retrieve: trophic, trophic.raw
occurrence}

\item{info}{the type of info return}

\item{nb.cpu}{number of CPU to use}

\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{metaweb}{a \code{data.frame} with all species interactions listed}

\item{folder.gbif}{a \code{character}: folders in which gbif data are stored}

\item{folder.iucn.buffer}{a \code{character}, folder in which IUCN buffered range
are stored as raster files.}

\item{sampling.effort}{a \code{\link{sampling_effort}} object, generated by
\code{\link{calc_sampling_effort}} containing several layers of sampling
effort based on taxonomic group and the association between species and
the sampling effort layer}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{min.gbif}{a \code{numeric} (\emph{default} \code{25}), the minimum
number of gbif occurrences below which a species is filtered.}

\item{buffer.config}{a named \code{list} with taxonomic groups (Class, Order,
Family, Species) and associated distance that will be used to buffer IUCN
range. If several taxonomic groups are valid for a given species, the
lowest one will be prioritized. Buffer distances are given in km.}

\item{quantile.absence.certain}{a \code{numeric} (\emph{default}
\code{0.1}), between 0 and 1. The quantile of sampling effort below which
a species absence is considered as uncertain}

\item{prop.prey.certain}{a \code{numeric} (\emph{default} \code{0.8}),
between 0 and 1. The minimum proportion of prey occurrence considered as
certain for a raster cell to be included in the species dataset.}

\item{uncertain.value}{a \code{integer} (\code{0} or \code{1} ;
\emph{default} \code{0}). The value given to uncertain absences in cells
where more than \code{prop.prey.certain} prey are considered as certain.}
}
\description{
Class returned by \code{\link{prepare_dataset_gbif}} and
  \code{\link{prepare_dataset_iucn}}, that prepare a dataset for a trophic
  SDM, based on a gbif or IUCN workflow.
}
\details{
\describe{
  \item{gbif dataset (\code{prepare_gbif_dataset})}{ Workflow to prepare gbif
  dataset}
  \item{IUCN dataset (\code{prepare_iucn_dataset})}{ Workflow to prepare iucn
  dataset}
  }
}
\section{Slots}{

\describe{
\item{\code{summary.occurrence}}{a \code{data.frame} with the list of species and 
the total number of presence and absence in the occurrence dataset extracted
from gbif and independently of prey cell selection.}

\item{\code{summary.predator}}{a \code{data.frame} with a summary of the actual 
trophic dataset, i.e. list of species and the total number of presence 
and absence inside/outside IUCN range as well as additional information 
such as the number of prey.}

\item{\code{summary.prey}}{a \code{data.frame} with a summary of occurrence and 
prevalence for all prey associated to a predator dataset.}

\item{\code{file.occurrence.link}}{a named \code{vector} with a path to raw 
occurrence data for each species.}

\item{\code{file.trophic.link}}{a named \code{vector} with a path to filtered
trophic data for each species.}

\item{\code{metaweb.filtered}}{a \code{data.frame} with all species interactions 
that remain after filtering of species and prey.}

\item{\code{checklist.filtered}}{a \code{data.frame} with information on all species 
that remain after filtering of species.}

\item{\code{kept.species}}{a \code{character} vector with all species kept}

\item{\code{filtered.species}}{a named \code{list} vector with all species removed 
as well as the motivation of removal}

\item{\code{kept.prey}}{a named \code{list} with the prey kept for each species}

\item{\code{filtered.prey}}{a named \code{list} vector with the prey removed for 
each species as well as the motivation of removal}

\item{\code{param.raw}}{a \code{list} with information to regenerate the raw 
unfiltered dataset.}

\item{\code{param}}{a \code{list} with all parameters used to generate the dataset}

\item{\code{datatype}}{a \code{character} determining the origin of the dataset:
either 'gbif' or 'iucn'.}

\item{\code{project.name}}{a \code{character} indicating the folder in which
logfiles and data may be written}

\item{\code{sampling.effort}}{a \code{sampling_effort} object describing the sampling
effort associated to all taxa.}

\item{\code{data.mask}}{a \code{SpatRaster} with the extent and the grid used 
in the project}
}}

\examples{

showClass("trophic_dataset")
}
\author{
Remi Patin
}
