% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_trophic_dataset.R, R/load_data.R,
%   R/prepare_dataset.R
\docType{class}
\name{trophic_dataset}
\alias{trophic_dataset}
\alias{trophic_dataset-class}
\alias{show,trophic_dataset-method}
\alias{subsample}
\alias{subsample,trophic_dataset-method}
\alias{reset}
\alias{reset,trophic_dataset-method}
\alias{summary_trophic}
\alias{summary_trophic,trophic_dataset-method}
\alias{plot_uncertain}
\alias{plot_uncertain,trophic_dataset-method}
\alias{plot_trophic}
\alias{plot_trophic,trophic_dataset-method}
\alias{plot_dataset}
\alias{plot_dataset,trophic_dataset-method}
\alias{rasterize_uncertain}
\alias{rasterize_uncertain,trophic_dataset-method}
\alias{rasterize_prey_filtering}
\alias{rasterize_prey_filtering,trophic_dataset-method}
\alias{load_data}
\alias{load_data,trophic_dataset-method}
\alias{prepare_dataset}
\title{Dataset for running trophic SDM}
\usage{
\S4method{show}{trophic_dataset}(object)

subsample(x, ...)

\S4method{subsample}{trophic_dataset}(x, param.subsampling)

reset(x, ...)

\S4method{reset}{trophic_dataset}(x, type = "full")

summary_trophic(x, ...)

\S4method{summary_trophic}{trophic_dataset}(x, info = "trophic_dataset")

plot_uncertain(x, ...)

\S4method{plot_uncertain}{trophic_dataset}(x, nb.cpu = 1)

plot_trophic(x, ...)

\S4method{plot_trophic}{trophic_dataset}(x, nb.cpu = 1)

plot_dataset(x, ...)

\S4method{plot_dataset}{trophic_dataset}(x, type, nb.cpu = 1)

rasterize_uncertain(x, ...)

\S4method{rasterize_uncertain}{trophic_dataset}(x, raster.by = "Class")

rasterize_prey_filtering(x, ...)

\S4method{rasterize_prey_filtering}{trophic_dataset}(x, raster.by = "Class")

load_data(x, ...)

\S4method{load_data}{trophic_dataset}(x, SpeciesName, Code, type = "trophic")

prepare_dataset(
  checklist,
  metaweb,
  use.gbif = TRUE,
  param.gbif,
  param.subsampling,
  folder.iucn,
  data.mask,
  project.name,
  nb.cpu,
  overwrite.occurrence = FALSE,
  overwrite.trophic = FALSE,
  subsample.assemble,
  ...
)
}
\arguments{
\item{object}{an object of class \code{trophic_dataset}}

\item{x}{an object of class \code{trophic_dataset}}

\item{type}{a \code{character}, the type of data to retrieve: trophic, trophic.raw
occurrence}

\item{info}{the type of info return}

\item{nb.cpu}{number of CPU to use}

\item{SpeciesName}{a \code{character}, a species name to retrieve data or 
information}

\item{Code}{a \code{character}, a species code to retrieve data or information}

\item{checklist}{a \code{data.frame} with information on all species of interest}

\item{metaweb}{a \code{data.frame} with all species interactions listed}

\item{folder.iucn}{a \code{character}, folder in which IUCN buffered range
are stored as raster files.}

\item{data.mask}{a \code{SpatRaster} with the extent and the grid used 
in the project}

\item{project.name}{a \code{character} indicating the folder in which logfiles 
and data may be written}

\item{sampling.effort}{a \code{\link{sampling_effort}} object, generated by
\code{\link{calc_sampling_effort}} containing several layers of sampling
effort based on taxonomic group and the association between species and
the sampling effort layer}

\item{min.gbif}{a \code{numeric} (\emph{default} \code{25}), the minimum
number of gbif occurrences below which a species is filtered.}

\item{folder.iucn.buffer}{a \code{character}, folder in which IUCN buffered range
are stored as raster files.}
}
\description{
Class returned by \code{\link{prepare_dataset}} and
  \code{\link{prepare_dataset_iucn}}, that prepare a dataset for a trophic
  SDM, based on a gbif or IUCN workflow.
}
\section{Slots}{

\describe{
\item{\code{summary}}{a \code{\link{trophic_summary}} with summary information on the workflow}

\item{\code{summary.raw}}{a \code{\link{trophic_summary}} with raw summary
information on the workflow (before filtering and subsampling)}

\item{\code{files}}{a \code{\link{trophic_files}} with links to data files}

\item{\code{param}}{a \code{\link{trophic_param}} with the parameter used in the 
workflow}

\item{\code{species}}{a \code{\link{trophic_species}} with information on all species 
that remains after filtering of species.}

\item{\code{project.name}}{a \code{character} indicating the folder in which
logfiles and data may be written}

\item{\code{data.mask}}{a \code{SpatRaster} with the extent and the grid used 
in the project}
}}

\examples{

showClass("trophic_dataset")
}
\author{
Remi Patin
}
